Move CollectionMap (a detail of Database impl) to Database folder
